refactor: shorten dir name
