+++ moqt-server/src/modules/message_handlers.rs
@@ -1,3 +1,3 @@
-pub(crate) mod control_message_handler;
-pub(crate) mod object_stream_handler;
-pub(crate) mod stream_header_handler;
+pub(crate) mod control_message;
+pub(crate) mod object_stream;
+pub(crate) mod stream_header;
