+++ moqt-server/src/modules/message_handlers/control_message/handlers/subscribe_namespace_handler.rs
