+++ moqt-server/src/modules/message_handlers/control_message/handlers/subscribe_ok_handler.rs
@@ -240,7 +240,7 @@ mod success {
 
 #[cfg(test)]
 mod failure {
-    use crate::modules::message_handlers::control_message_handler::handlers::subscribe_ok_handler::subscribe_ok_handler;
+    use crate::modules::message_handlers::control_message::handlers::subscribe_ok_handler::subscribe_ok_handler;
     use crate::modules::pubsub_relation_manager::{
         commands::PubSubRelationCommand, manager::pubsub_relation_manager,
         wrapper::PubSubRelationManagerWrapper,
