+++ moqt-server/src/modules/message_handlers/control_message/server_processes/subscribe_ok_message.rs
@@ -8,7 +8,7 @@ use moqt_core::{
 };
 
 use crate::modules::{
-    message_handlers::control_message_handler::handlers::subscribe_ok_handler::subscribe_ok_handler,
+    message_handlers::control_message::handlers::subscribe_ok_handler::subscribe_ok_handler,
     moqt_client::MOQTClient,
 };
